Move fileIdSelection.ts into its own folder
