+++ apps/desktop/src/components/BranchPreview.svelte
@@ -7,8 +7,8 @@
 	import ScrollableContainer from '$components/ScrollableContainer.svelte';
 	import { transformAnyCommit } from '$lib/commits/transformers';
 	import { getForge } from '$lib/forge/interface/forge';
+	import { FileIdSelection } from '$lib/selection/fileIdSelection';
 	import { SETTINGS, type Settings } from '$lib/settings/userSettings';
-	import { FileIdSelection } from '$lib/vbranches/fileIdSelection';
 	import { BranchData } from '$lib/vbranches/types';
 	import { getContext, getContextStoreBySymbol } from '@gitbutler/shared/context';
 	import Line from '@gitbutler/ui/commitLines/Line.svelte';
