+++ apps/desktop/src/components/FileListItemWrapper.svelte
@@ -5,9 +5,9 @@
 	import { getCommitStore } from '$lib/commits/contexts';
 	import { draggableChips, type DraggableConfig } from '$lib/dragging/draggable';
 	import { FileDropData } from '$lib/dragging/draggables';
+	import { FileIdSelection } from '$lib/selection/fileIdSelection';
 	import { itemsSatisfy } from '$lib/utils/array';
 	import { computeFileStatus } from '$lib/utils/fileStatus';
-	import { FileIdSelection } from '$lib/vbranches/fileIdSelection';
 	import { getLockText } from '$lib/vbranches/tooltip';
 	import { BranchStack, type AnyFile, LocalFile } from '$lib/vbranches/types';
 	import { getContext, maybeGetContextStore } from '@gitbutler/shared/context';
