+++ apps/desktop/src/lib/selection/fileIdSelection.ts
