+++ apps/desktop/src/lib/utils/selection.ts
@@ -3,8 +3,8 @@
  */
 import { getSelectionDirection } from './getSelectionDirection';
 import { KeyName } from './hotkeys';
-import { stringifyKey, unstringifyFileKey } from '$lib/vbranches/fileIdSelection';
-import type { FileIdSelection } from '$lib/vbranches/fileIdSelection';
+import { stringifyKey, unstringifyFileKey } from '$lib/selection/fileIdSelection';
+import type { FileIdSelection } from '$lib/selection/fileIdSelection';
 import type { AnyFile } from '$lib/vbranches/types';
 
 function getFile(files: AnyFile[], id: string): AnyFile | undefined {
